chore(CODEOWNERS): Remove global ownership and transfer it
In the context of QA-1017, the original motivation was to move away from deprecated
dependabot reviewers but not necessarily enable code reviewers feature across all files. This commit
amends the original work so that only dependency update related PRs are responsibility of the
client-dependabot-reviewers team

Furthermore, transfer the ownership for the update from SRE team to
Client team. It is unclear why we had such a inconsistency between the
two GH actions repos but are aligning them now.

Ticket: QA-1017
Signed-off-by: Lluis Campos <lluis.campos@northern.tech>